Refactor Viewed Router and DAO
